+++ apps/admin/src/features/heroImagesManager/components/HeroImageEditCard/EditHeroImageButton.tsx
@@ -3,7 +3,7 @@ import { PencilIcon } from 'lucide-react'
 
 import { HeroImageForm } from '../HeroImageForm'
 
-import { useModal } from '~/hooks/use-modal'
+import { useModal } from '~/hooks/useModal'
 import { MESSAGES } from '../../constant/constants'
 
 import type { CarouselResponse } from '~/apis/community/requests'
