+++ apps/admin/src/shared/components/Post/Board.tsx
@@ -11,7 +11,7 @@ import {
 } from 'lucide-react'
 
 import { EDIT_POST_PATH_MAP, type PostCategory } from '~/constants/path'
-import { useModal } from '~/hooks/use-modal'
+import { useModal } from '~/hooks/useModal'
 import { MESSAGES } from '~/shared/constant/post.constants'
 import { useDeletePost } from '~/shared/hooks/Post'
 import { extractFileName } from '~/shared/utils/utils'
